Refactor promotion handling to use Double for cocoporc across controllers and model, ensuring consistent data type for discount percentages. Update index.jsp for improved display and navigation functionality with new onclick events for feature cards.

+++ src/main/java/com/example/Controllers/ExportPromotionsController.java
@@ -162,9 +162,9 @@ public class ExportPromotionsController extends HttpServlet {
                     promotion.setConcdesc("");
                 }
                 try {
-                    promotion.setCocoporc(rs.getInt("COCOPORC"));
+                    promotion.setCocoporc(rs.getDouble("COCOPORC"));
                 } catch (Exception e) {
-                    promotion.setCocoporc(0);
+                    promotion.setCocoporc(0.0);
                 }
                 try {
                     promotion.setCocotiap(rs.getInt("COCOTIAP"));
